Converting to gradle project
Added build.gradle.
Removed eclipse project.
Removed all jars except txns (and updated this to alpha.4).  Others will be pulled in by gradle.
Upped to java-client alpha.3 and fixed some Java code to be compatible.
Removed "transactions" and "doc_operations_sdk3" projects, moved their
logic into TAF, in src dir.
Gradle task testrunner is running jython
