+++ .gitignore
@@ -7,4 +7,5 @@
 **/.idea
 **/logs
 **/settings
-bulk_doc_operations/target
+build
+.gradle
