--- .pydevproject
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<?eclipse-pydev version="1.0"?><pydev_project>
-<pydev_pathproperty name="org.python.pydev.PROJECT_SOURCE_PATH">
-    <path>/${PROJECT_DIR_NAME}/bulk_doc_operations.jar</path>
-    <path>/${PROJECT_DIR_NAME}/jsch-0.1.54.jar</path>
-    <path>/${PROJECT_DIR_NAME}/lib</path>
-    <path>/${PROJECT_DIR_NAME}/connections</path>
-    <path>/${PROJECT_DIR_NAME}/pytests</path>
-    <path>/${PROJECT_DIR_NAME}/scripts</path>
-    <path>/${PROJECT_DIR_NAME}/couchbase_utils</path>
-    <path>/${PROJECT_DIR_NAME}/platform_utils</path>
-    <path>/${PROJECT_DIR_NAME}/sdk3.0/core-io-2.0.0-alpha.2.jar</path>
-    <path>/${PROJECT_DIR_NAME}/sdk3.0/java-client-3.0.0-alpha.2.jar</path>
-    <path>/${PROJECT_DIR_NAME}/sdk3.0/opentracing-api-0.31.0.jar</path>
-    <path>/${PROJECT_DIR_NAME}/sdk3.0/reactive-streams-1.0.2.jar</path>
-    <path>/${PROJECT_DIR_NAME}/sdk3.0/reactor-core-3.2.8.RELEASE.jar</path>
-    <path>/${PROJECT_DIR_NAME}/sdk3.0/java-client-3.0.0-alpha.2-javadoc.jar</path>
-    <path>/${PROJECT_DIR_NAME}/doc_operations_sdk3.jar</path>
-    <path>/${PROJECT_DIR_NAME}</path>
-    <path>/TAF_Couchbaselabs/Transaction.jar</path>
-    <path>/TAF_Couchbaselabs/TransactionsExample-1.0.0-all.jar</path>
-</pydev_pathproperty>
-    <pydev_property name="org.python.pydev.PYTHON_PROJECT_VERSION">jython interpreter</pydev_property>
-    <pydev_property name="org.python.pydev.PYTHON_PROJECT_INTERPRETER">Default</pydev_property>
-</pydev_project>
