--- transactions/src/main/java/com/couchbase/test/transactions/App.java
@@ -1,13 +1,0 @@
-package com.couchbase.test.transactions;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
